sstables: writer: Track first and last clustering positions
Part of ongoing effort to write first and last clustering positions,
(which belong to first and last partition keys respectively) into
Scylla.DB component.

Commit 9bcad9ffa89 introduced procedure to read it from first and
last key, but it's problematic due to amount of resources and time
we have to devote for doing it. May significantly slow down sstable
loading and cause large allocation issues when reading in
reverse.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>