sstables: Introduce clustering position metadata
This new metadata will be written to Scylla component.

The structure will store the first and last clustering positions,
belonging to first and last partition keys respectively.

This metadata is intended to be used in the context of sstable
runs, which may have a large partition split across several
SSTables.

Today, we're reading from the actual Datafile, but that caused
a few complications, so we're going to replace the procedure
by just reading this new metadata instead.

Unlike the min/max metadata from Statistics.db, position for
partition tombstone is not accounted for, because sstable run
doesn't break the invariant in which a tombstone can be found
in any fragment storing a particular partition.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>