sstables: Write clustering position metadata into SSTable
The metadata will now be persisted into Scylla-specific SSTable
component.

We don't need to worry about backward compatibility because no
compaction strategy is splitting large partitions yet.

We can also pick safe defaults if the metadata is not present:
{ position_in_partition::before_all_clustered_rows(),
  position_in_partition::after_all_clustered_rows() }.
Meaning the full clustering range is contained in the file,
and therefore cannot be dismissed when querying a particular
partition, and also it cannot co-exist in a sstable run with
another fragment containing the same partition, as they'll
overlap at clustering level.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>