table: Don't create FS directories explicitly
The (newly introduced) make_directory_for_column_family() goes directly
to the filesystem and makes directories with specific layout. For object
storage this is not going to work, so instead, move the dir making code
into sstables manager and call it via .initialize_storage().

For now the initialize_storage() accepts the list of paths to mkdir,
firther patches will make this object-storage-aware too.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>