view: row_locker: lock_holder: add lock_{partition,row} functions
We want the lock_holder state to exacty represent
the locking state of each lock so if it is destroyed
prematurely (e.g. by a timeout) it will unlock only
the particular locks it acquired.

This patch moves the creation of the lock_holder
upfront, with null partition and row keys.

Then lock_partition and lock_row respectively populate
the state for the locked partition or row so it can
be cleaned up properly when the lock_holder is destroyed.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>