generic_server: split connection::process() into main part and error handling
This makes it easier to handle the "finally" for coroutinization.

Note that connection::process() is a virtual function, but
the split doesn't change anything. If anything overrides it, then
the contents of both functions will be taken over by the derived
class, since the entry point is the virtual function.