generic_server: coroutinize connection::process{,_main}()
process_main() is relatively straightforward to coroutinize. Since
process_request() usually blocks (even if reading the request
doesn't block, the execution stage will), we save an allocation
for the then_wrapped() continuation.

process() is more difficult due to the behavior of seastar::finally()
which doesn't have a direct equivalent in C++. We add a merge_exception()
helper to merge exceptions encountered during the cleanup into the
process_main()'s exception, giving priority to the first exception
to be seen.